+++ miniconf/src/json_core.rs
@@ -1,12 +1,14 @@
-use crate::{Error, IntoKeys, TreeDeserialize, TreeSerialize};
 use serde_json_core::{de, ser};
+
+use crate::{Error, IntoKeys, Path, TreeDeserialize, TreeSerialize};
 
 /// `TreeSerialize`/`TreeDeserialize` with "JSON and `/`".
 ///
 /// Access items with `'/'` as path separator and JSON (from `serde-json-core`)
 /// as serialization/deserialization payload format.
 ///
-/// Paths used here are reciprocal to `TreeKey::path(..., "/")`/`TreeKey::iter_paths("/")`.
+/// Paths used here are reciprocal to `TreeKey::lookup::<Path<_, '/'>, _>(...)`/
+/// `TreeKey::nodes::<Path<_, '/'>>()`.
 ///
 /// ```
 /// use miniconf::{JsonCoreSlash, Tree};
@@ -71,11 +73,11 @@ impl<'de, T: TreeSerialize<Y> + TreeDeserialize<'de, Y> + ?Sized, const Y: usize
     JsonCoreSlash<'de, Y> for T
 {
     fn set_json(&mut self, path: &str, data: &'de [u8]) -> Result<usize, Error<de::Error>> {
-        self.set_json_by_key(path.split('/').skip(1), data)
+        self.set_json_by_key(&Path::<_, '/'>::from(path), data)
     }
 
     fn get_json(&self, path: &str, data: &mut [u8]) -> Result<usize, Error<ser::Error>> {
-        self.get_json_by_key(path.split('/').skip(1), data)
+        self.get_json_by_key(&Path::<_, '/'>::from(path), data)
     }
 
     fn set_json_by_key<K: IntoKeys>(