+++ miniconf/tests/embedded/src/main.rs
@@ -9,7 +9,9 @@ use cortex_m_rt::entry;
 use cortex_m_semihosting::{debug, hprintln};
 
 use crosstrait::{register, Cast};
-use miniconf::{self, Tree, TreeAny};
+use miniconf::{
+    self, IntoKeys, JsonCoreSlash, JsonPath, Node, Packed, Path, Tree, TreeAny, TreeKey,
+};
 
 use core::ops::{AddAssign, SubAssign};
 register! { i32 => dyn AddAssign<i32> }
@@ -46,11 +48,20 @@ fn main() -> ! {
     assert_eq!(a, 3 + 5);
 
     let mut s = Settings::default();
-    s.i[1].val = 3;
 
-    // let key = miniconf::Packed::new_from_lsb(0b1_01_01_0).unwrap();
-    let key = miniconf::JsonPath::from(".i[1].val");
-    let any = s.mut_any_by_key(key).unwrap();
+    let path = Path::<_, '/'>::from("/i/1/val");
+    s.set_json(&path, b"3").unwrap();
+
+    let (packed, node) = Settings::transcode::<Packed, _>(&path).unwrap();
+    assert_eq!(packed.into_lsb().get(), 0b1_01_01_0);
+    assert_eq!(node, Node::leaf(3));
+
+    let mut buf = [0; 10];
+    let len = s.get_json_by_key(packed, &mut buf).unwrap();
+    assert_eq!(&buf[..len], b"3");
+
+    let key = JsonPath::from(".i[1].val");
+    let any = s.mut_any_by_key(key.into_keys()).unwrap();
 
     let val: &mut dyn AddAssign<i32> = any.cast().unwrap();
     *val += 5;