+++ miniconf/tests/option.rs
@@ -1,4 +1,4 @@
-use miniconf::{Error, JsonCoreSlash, Traversal, Tree, TreeKey};
+use miniconf::{Error, JsonCoreSlash, Path, Traversal, Tree, TreeKey};
 
 #[derive(PartialEq, Debug, Clone, Default, Tree)]
 struct Inner {
@@ -13,8 +13,8 @@ struct Settings {
 
 #[test]
 fn just_option() {
-    let mut it = Option::<u32>::iter_paths::<String>("/").count();
-    assert_eq!(it.next(), Some(Ok("".into())));
+    let mut it = Option::<u32>::nodes::<Path<String, '/'>>().exact_size();
+    assert_eq!(it.next().unwrap().unwrap().0.as_str(), "");
     assert_eq!(it.next(), None);
 }
 
@@ -60,14 +60,16 @@ fn option_iterate_some_none() {
 
     // When the value is None, it will still be iterated over as a topic but may not exist at runtime.
     settings.value.take();
-    let mut iterator = Settings::iter_paths::<String>("/").count();
-    assert_eq!(iterator.next(), Some(Ok("/value/data".into())));
+    let mut iterator = Settings::nodes::<Path<String, '/'>>().exact_size();
+    let (path, _node) = iterator.next().unwrap().unwrap();
+    assert_eq!(path.as_str(), "/value/data");
     assert!(iterator.next().is_none());
 
     // When the value is Some, it should be iterated over.
     settings.value.replace(Inner { data: 5 });
-    let mut iterator = Settings::iter_paths::<String>("/").count();
-    assert_eq!(iterator.next(), Some(Ok("/value/data".into())));
+    let mut iterator = Settings::nodes::<Path<String, '/'>>().exact_size();
+    let (path, _node) = iterator.next().unwrap().unwrap();
+    assert_eq!(path.as_str(), "/value/data");
     assert_eq!(iterator.next(), None);
 }
 
@@ -81,15 +83,17 @@ fn option_test_normal_option() {
     let mut s = S::default();
     assert!(s.data.is_none());
 
-    let mut iterator = S::iter_paths::<String>("/").count();
-    assert_eq!(iterator.next(), Some(Ok("/data".into())));
+    let mut iterator = S::nodes::<Path<String, '/'>>().exact_size();
+    let (path, _node) = iterator.next().unwrap().unwrap();
+    assert_eq!(path.as_str(), "/data");
     assert!(iterator.next().is_none());
 
     s.set_json("/data", b"7").unwrap();
     assert_eq!(s.data, Some(7));
 
-    let mut iterator = S::iter_paths::<String>("/").count();
-    assert_eq!(iterator.next(), Some(Ok("/data".into())));
+    let mut iterator = S::nodes::<Path<String, '/'>>().exact_size();
+    let (path, _node) = iterator.next().unwrap().unwrap();
+    assert_eq!(path.as_str(), "/data");
     assert!(iterator.next().is_none());
 
     s.set_json("/data", b"null").unwrap();
@@ -107,8 +111,9 @@ fn option_test_defer_option() {
     let mut s = S::default();
     assert!(s.data.is_none());
 
-    let mut iterator = S::iter_paths::<String>("/").count();
-    assert_eq!(iterator.next(), Some(Ok("/data".into())));
+    let mut iterator = S::nodes::<Path<String, '/'>>().exact_size();
+    let (path, _node) = iterator.next().unwrap().unwrap();
+    assert_eq!(path.as_str(), "/data");
     assert!(iterator.next().is_none());
 
     assert!(s.set_json("/data", b"7").is_err());
@@ -116,8 +121,9 @@ fn option_test_defer_option() {
     s.set_json("/data", b"7").unwrap();
     assert_eq!(s.data, Some(7));
 
-    let mut iterator = S::iter_paths::<String>("/").count();
-    assert_eq!(iterator.next(), Some(Ok("/data".into())));
+    let mut iterator = S::nodes::<Path<String, '/'>>().exact_size();
+    let (path, _node) = iterator.next().unwrap().unwrap();
+    assert_eq!(path.as_str(), "/data");
     assert!(iterator.next().is_none());
 
     assert!(s.set_json("/data", b"null").is_err());